feat: add tags parameter to create_project
