+++ src/kili/adapters/kili_api_gateway/project/__init__.py
@@ -1,7 +1,8 @@
 """Mixin extending Kili API Gateway class with Project related operations."""
 
 
-from typing import Dict, List
+import json
+from typing import Dict, List, Optional
 
 from kili.adapters.kili_api_gateway.helpers.queries import fragment_builder
 from kili.adapters.kili_api_gateway.project.operations import get_project_query
@@ -9,13 +10,14 @@ from kili.domain.project import ProjectId
 from kili.exceptions import NotFound
 
 from ..base import BaseOperationMixin
+from .operations import GQL_CREATE_PROJECT
 
 
 class ProjectOperationMixin(BaseOperationMixin):
-    """Mixin extending Kili API Gateway class with Assets related operations."""
+    """Mixin extending Kili API Gateway class with Project related operations."""
 
     def get_project(self, project_id: ProjectId, fields: List[str]) -> Dict:
-        """List assets with given options."""
+        """Get a project."""
         fragment = fragment_builder(fields)
         query = get_project_query(fragment)
         result = self.graphql_client.execute(
@@ -28,3 +30,24 @@ class ProjectOperationMixin(BaseOperationMixin):
                 " to it."
             )
         return projects[0]
+
+    def create_project(
+        self,
+        input_type: str,
+        json_interface: dict,
+        title: str,
+        description: str,
+        project_type: Optional[str],
+    ) -> ProjectId:
+        """Create a project."""
+        variables = {
+            "data": {
+                "description": description,
+                "inputType": input_type,
+                "jsonInterface": json.dumps(json_interface),
+                "projectType": project_type,
+                "title": title,
+            }
+        }
+        result = self.graphql_client.execute(GQL_CREATE_PROJECT, variables)
+        return ProjectId(result["id"])
