+++ src/kili/client.py
@@ -39,6 +39,7 @@ from kili.exceptions import AuthenticationFailed, UserNotFoundError
 from kili.presentation.client.asset import AssetClientMethods
 from kili.presentation.client.internal import InternalClientMethods
 from kili.presentation.client.issue import IssueClientMethods
+from kili.presentation.client.project import ProjectClientMethods
 from kili.presentation.client.tag import TagClientMethods
 from kili.utils.logcontext import LogContext, log_call
 
@@ -81,6 +82,7 @@ class Kili(  # pylint: disable=too-many-ancestors,too-many-instance-attributes
     IssueClientMethods,
     AssetClientMethods,
     TagClientMethods,
+    ProjectClientMethods,
 ):
     """Kili Client."""
