+++ src/kili/entrypoints/mutations/project/helpers.py
@@ -8,9 +8,11 @@ def verify_argument_ranges(
     consensus_tot_coverage: Optional[int],
     min_consensus_size: Optional[int],
     review_coverage: Optional[int],
-):
-    """Ensure that all arguments are set in a correct range or raise
-    a ValueError otherwise."""
+) -> None:
+    """Ensure that all arguments are set in a correct range.
+
+    Raise a ValueError otherwise.
+    """
     if consensus_tot_coverage is not None and (
         consensus_tot_coverage < 0 or consensus_tot_coverage > 100
     ):
